requirements.txt
Flask==3.1.0
flask_mail==0.10.0
flask_sqlalchemy==3.1.1
joblib==1.3.2
numpy==1.23.5
numpy==1.24.3
pandas==2.2.3
Werkzeug==3.1.3
psycopg2-binary
python-dotenv